+++ include/viua/cg/bytecode/instructions.h
@@ -94,7 +94,7 @@ namespace cg {
         byte* opvat(byte*, int_op, int_op, int_op);
         byte* opvlen(byte*, int_op, int_op);
 
-        byte* opnot(byte*, int_op);
+        byte* opnot(byte*, int_op, int_op);
         byte* opand(byte*, int_op, int_op, int_op);
         byte* opor(byte*, int_op, int_op, int_op);